# transactions/src/main/resources/api.yml
openapi: '3.0.0'

info:
  version: 1.0.0
  title: Transacciones
  description: Gestión de transacciones o movimientos bancarios

servers:
  - url: http://localhost:8584
tags:
  - name: transactions
    description: Manejo de transacciones o movimientos bancarios
paths:
  /transactions:
    post:
      tags:
        - transactions
      summary: Save Transaction
      operationId: saveTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionReq'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionRes'
    get:
      tags:
        - transactions
      summary: Get transactions
      operationId: getTransactions
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionRes'
  /transactions/{id}:
    get:
      tags:
        - transactions
      summary: Get transaction by id
      operationId: getTransactionById
      parameters:
        - name: id
          in: path
          description: Unique id of an transaction
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionRes'
    put:
      tags:
        - transactions
      summary: Update Transaction
      operationId: updateTransaction
      parameters:
        - name: id
          in: path
          description: Unique id of a transaction
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionReq'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionRes'
  /transactions/{clientId}/{year}/{month}:
    get:
      tags:
        - transactions
      summary: Get transaction by client, year and month.
      operationId: getClientTransactions
      parameters:
        - name: clientId
          in: path
          description: Unique client id
          required: true
          schema:
            type: string
        - name: year
          in: path
          description: Unique year
          required: true
          schema:
            type: integer
        - name: month
          in: path
          description: Unique month
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionRes'

components:
  schemas:
    TransactionReq:
      type: object
      description: Data element for transaction
      properties:
        client:
          type: string
        type:
          type: string
          enum:
            - DEPOSITO
            - RETIRO
            - PAGO
        amount:
          type: number
        origin:
          type: string
        destiny:
          type: string
      required:
        - client
        - type
        - amount
        - origin
        - destiny
    TransactionRes:
      type: object
      description: Data element for transaction
      properties:
        transactionId:
          type: string
        client:
          type: string
        type:
          type: string
        amount:
          type: number
        origin:
          type: string
        destiny:
          type: string